+++ src/view/sorting-view.js
@@ -1,25 +1,11 @@
-import { createElement } from '../service/render.js';
+import AbstractView from './abstract-view.js';
 
 export default class SortingView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
   get template() {
     return `<ul class="sort">
     <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
     <li><a href="#" class="sort__button">Sort by date</a></li>
     <li><a href="#" class="sort__button">Sort by rating</a></li>
     </ul>`;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
